:sparkles: Wrap with native `<Suspense />`
